app/requirements.txt
streamlit
plotly
pandas
snowflake-connector-python[pandas]
python-dotenv